fix: Search for categories of product type moved to Navbar.
