+++ src/components/navbar/Navbar.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import {
   Collapse,
   Navbar,
@@ -16,6 +16,9 @@ import { Link } from "react-router-dom";
 import "../../styles/Navbar.css";
 import { Input, Button } from "semantic-ui-react";
 import { TextField } from "@material-ui/core";
+import { Drawer } from "../../components/menu/index";
+import productManager from "../../modules/productManager";
+import ChevronRightIcon from "@material-ui/icons/ChevronRight";
 
 const Example = ({
   navArray = defaultArray,
@@ -28,8 +31,69 @@ const Example = ({
   handleSubmit,
 }) => {
   const [isOpen, setIsOpen] = useState(false);
+  const [drawerOpen, setDrawerOpen] = useState(false);
+  const [productTypes, setProductTypes] = useState([]);
+  const [products, setProducts] = useState([]);
 
   const toggle = () => setIsOpen(!isOpen);
+  const toggleMenu = () => {
+    setDrawerOpen(!drawerOpen);
+  };
+
+  // Function that maps the producttype id's to the products table where the product_type_id is, and counts the number of products under the specific producttype.
+  // Then creates a new custom object inside an array that is set to state so the data is easily iterable.
+  // This function renders data for the Drawer.
+  const productCategories = async () => {
+    try {
+      const getAllProductTypes = await productManager.getProductTypes();
+      const productTypeMap = getAllProductTypes.map((productType) => {
+        const productTypeId = productType.id;
+        return productTypeId;
+      });
+      let promises = [];
+      let productTypeArray = [];
+      let productsArray = [];
+      let newObj = {};
+      productTypeMap.forEach((id) => {
+        promises.push(productManager.getProductsByProductType(id));
+      });
+      Promise.all(promises)
+        .then((data) => {
+          data.forEach((productType) => {
+            const len = productType.length;
+            const mapName = productType.map(
+              (product) => product.product_type.name
+            );
+            const name = mapName[0];
+            const mapId = productType.map((product) => product.product_type.id);
+            const id = mapId[0];
+            productsArray.push(productType);
+            newObj = {
+              id: id,
+              name: name,
+              count: len,
+              products: productsArray.filter(
+                (prod) => productType.id === prod.product_type_id
+              ),
+            };
+            productTypeArray.push(newObj);
+          });
+          setProductTypes(productTypeArray);
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    } catch (err) {
+      console.log(err);
+    }
+  };
+
+  useEffect(() => {
+    productCategories();
+    productManager.getHomeList().then((arr) => {
+      setProducts(arr);
+    });
+  }, []);
 
   return (
     <div>
@@ -82,6 +146,63 @@ const Example = ({
               }
             })}
           </Nav>
+
+          <NavItem className="navbar-item-link">
+            <div className="product-category-container">
+              <div className="category-items">
+                <Link onClick={() => toggleMenu()}>
+                  Search Products By Categories
+                </Link>
+                <Drawer
+                  position="left"
+                  isOpen={drawerOpen}
+                  close={() => toggleMenu()}
+                  drawerInfo={productTypes.map((productType) => {
+                    const pArray = productType.products;
+                    return (
+                      <>
+                        <div className="producttypes-items-container">
+                          <Link
+                            onClick={() => toggleMenu()}
+                            key={productType.id}
+                            to={`/products/category/${productType.id}`}
+                          >
+                            <span className="product-name-on-list">
+                              <strong>{productType.name}</strong> - (
+                              {productType.count}) Items
+                            </span>
+                            <ChevronRightIcon />
+                          </Link>
+                          <div className="top-sellers">Top Sellers:</div>
+                          <ul className="top-products-container">
+                            {pArray.map((products) => {
+                              return products.slice(0, 3).map((product) => {
+                                if (
+                                  product.product_type_id === productType.id
+                                ) {
+                                  return (
+                                    <li className="product-items-list">
+                                      <Link
+                                        onClick={() => toggleMenu()}
+                                        to={`/products/${product.id}`}
+                                        key={product.id}
+                                      >
+                                        {product.title}
+                                      </Link>
+                                    </li>
+                                  );
+                                }
+                              });
+                            })}
+                          </ul>
+                        </div>
+                      </>
+                    );
+                  })}
+                />
+              </div>
+            </div>
+          </NavItem>
           <Nav navbar>
             <form>
               <NavItem className="navbar-item-link">
@@ -167,7 +288,7 @@ const defaultArray = [
 export default Example;
 
 // {
-  /* <NavItem>
+/* <NavItem>
               <Input
                 placeholder="Search..."
                 action={
