+++ src/pages/Router.js
@@ -9,13 +9,13 @@ import {
 import { DLHOME, Profile, Reports } from "./index";
 import { Navbar } from "../components";
 import "../styles/Global.css";
-import {ProductFormMaster} from "./products/";
+import { ProductFormMaster } from "./products/";
 import { HomePage } from "./home/index";
 import { ProductDetails } from "./products/index";
 import { Register, Login } from "../pages/users/index";
 import MyProducts from "./products/MyProducts";
 import { MyCart, Checkout } from "./orders/index";
-import { ProductType } from "./products/index"
+import { ProductType } from "./products/index";
 import SearchForm from "../components/form/searchForm";
 
 const Routes = (props) => {
@@ -63,7 +63,7 @@ const Routes = (props) => {
                 {
                   title: <i className="shopping cart icon"></i>,
                   route: "/mycart",
-              },
+                },
               ]
             : []
         }
@@ -90,7 +90,7 @@ const Routes = (props) => {
         />
 
         {/* <React.Fragment> */}
-          <div className="body-container">
+        <div className="body-container">
           <Route
             exact
             path="/register"
@@ -130,7 +130,7 @@ const Routes = (props) => {
             render={(props) =>
               hasUser ? <MyProducts {...props} /> : <Redirect to="/" />
             }
-            />
+          />
 
           {/* this will route will filter HomePage to product types */}
           <Route
@@ -138,15 +138,17 @@ const Routes = (props) => {
             // path="/products/:productTypeName([\w ]+)"
             path="/products/category/:productTypeId(\d+)"
             render={(props) =>
-              hasUser ? <ProductType
-                productTypeId={parseInt(props.match.params.productTypeId)}
-                // productTypeName={props.match.params.productTypeName}
-                {...props}
-              />
-                : <Redirect to="/" />
+              hasUser ? (
+                <ProductType
+                  productTypeId={parseInt(props.match.params.productTypeId)}
+                  // productTypeName={props.match.params.productTypeName}
+                  {...props}
+                />
+              ) : (
+                <Redirect to="/" />
+              )
             }
           />
-
 
           {/* this will route to a product detail page */}
           <Route
@@ -227,8 +229,8 @@ const Routes = (props) => {
             render={(props) =>
               hasUser ? <Profile {...props} /> : <Redirect to="/" />
             }
-            />
-            </div>
+          />
+        </div>
         {/* </React.Fragment> */}
         {/* Will redirect to home page if page does not exist */}
         <Redirect to="/" />
